Conflicts:
	CZHV/Crouching Zombie Hidden Viking The Game/defaultlighting/defaultlighting.frag
	CZHV/Crouching Zombie Hidden Viking The Game/defaultlighting/defaultlighting.vert
	CZHV/Crouching Zombie Hidden Viking The Game/src/controller/InputManager.java
	CZHV/Crouching Zombie Hidden Viking The Game/src/view/renderer3D/core/Renderer3D.java

+++ CZHV/Crouching Zombie Hidden Viking The Game/src/controller/actions/ShootAction.java
@@ -43,7 +43,7 @@ public class ShootAction implements Action {
 		return false;
 	}
 
-	private boolean hitSuccess(int accuracy){
+	private boolean hitSuccess(float accuracy){
 		return Math.random() < accuracy;
 	}
